# openapi.yml
# Copilot-generiert: Kombinierte OpenAPI-Spezifikation aus meet-service.yml und registration-service.yml
# Zweck: API-Gateway-basierter Single-Entry-Point für alle CarLink Services
openapi: 3.1.0
info:
  title: CarMeet Combined API
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Meet-Service
  - url: http://localhost:3001
    description: Registration-Service
tags:
  - name: Venues
  - name: Meets
  - name: Users
  - name: Vehicles
  - name: Registrations
paths:
  /venues:
    get:
      tags: [Venues]
      summary: List venues
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Venue"
    post:
      tags: [Venues]
      summary: Create a venue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VenueCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
        "400":
          $ref: "#/components/responses/BadRequest"
  /venues/{venueId}:
    get:
      tags: [Venues]
      summary: Get a venue by id
      parameters:
        - $ref: "#/components/parameters/VenueId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Venues]
      summary: Update venue
      parameters:
        - $ref: "#/components/parameters/VenueId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VenueUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Venues]
      summary: Delete venue
      parameters:
        - $ref: "#/components/parameters/VenueId"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Venue is still used by one or more meets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /meets:
    get:
      tags: [Meets]
      summary: List meets (optionally filtered)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [PLANNED, CANCELLED, DONE]
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MeetResponse"
    post:
      tags: [Meets]
      summary: Create a meet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeetCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /meets/{meetId}:
    get:
      tags: [Meets]
      summary: Get meet by id
      parameters:
        - $ref: "#/components/parameters/MeetId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Meets]
      summary: Update meet (replace editable fields)
      parameters:
        - $ref: "#/components/parameters/MeetId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeetUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Meets]
      summary: Cancel a meet (soft delete)
      parameters:
        - $ref: "#/components/parameters/MeetId"
      responses:
        "200":
          description: Cancelled (status=CANCELLED)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /users:
    post:
      tags: [Users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /users/{userId}:
    get:
      tags: [Users]
      summary: Get user by id
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Users]
      summary: Update user
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [Users]
      summary: Delete user
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: User is still referenced by registrations or vehicles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /vehicles:
    get:
      tags: [Vehicles]
      summary: List vehicles (optional userId filter)
      parameters:
        - name: userId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vehicle"
    post:
      tags: [Vehicles]
      summary: Create vehicle (userId required in body)
      description: Create a vehicle record bound to a user. GET /vehicles is also supported.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleCreateWithUser"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /vehicles/{vehicleId}:
    get:
      tags: [Vehicles]
      summary: Get vehicle by id
      parameters:
        - $ref: "#/components/parameters/VehicleId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Vehicles]
      summary: Update vehicle
      parameters:
        - $ref: "#/components/parameters/VehicleId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Vehicles]
      summary: Delete vehicle
      parameters:
        - $ref: "#/components/parameters/VehicleId"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Vehicle is still referenced by one or more registrations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /registrations:
    get:
      tags: [Registrations]
      summary: List registrations (filter by meetId and/or userId)
      parameters:
        - name: meetId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: userId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: role
          in: query
          required: false
          schema:
            type: string
            enum: [PARTICIPANT, VISITOR]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Registration"
    post:
      tags: [Registrations]
      summary: Create registration for a meet (participant or visitor)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: User or Vehicle not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Duplicate registration (same userId + meetId)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /registrations/{registrationId}:
    get:
      tags: [Registrations]
      summary: Get registration by id
      parameters:
        - $ref: "#/components/parameters/RegistrationId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Registrations]
      summary: Update registration
      parameters:
        - $ref: "#/components/parameters/RegistrationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Registrations]
      summary: Cancel registration (soft delete)
      parameters:
        - $ref: "#/components/parameters/RegistrationId"
      responses:
        "200":
          description: Cancelled (status=CANCELLED)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
        "404":
          $ref: "#/components/responses/NotFound"
  /registrations/meets/{meetId}/count:
    get:
      tags: [Registrations]
      summary: Count registrations for a specific meet
      parameters:
        - name: meetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: role
          in: query
          required: false
          description: Optional role filter for participant/visitor counts
          schema:
            type: string
            enum: [PARTICIPANT, VISITOR]
      responses:
        "200":
          description: Registration count for meet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationCount"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  parameters:
    MeetId:
      name: meetId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    VenueId:
      name: venueId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    VehicleId:
      name: vehicleId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RegistrationId:
      name: registrationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [message]
      properties:
        message:
          type: string
        details:
          type: string
    Venue:
      type: object
      additionalProperties: false
      required: [venueId, name, city, country]
      properties:
        venueId:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
        street:
          type: string
        zip:
          type: string
        city:
          type: string
          minLength: 1
        country:
          type: string
          minLength: 1
        latitude:
          type: number
        longitude:
          type: number
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
    VenueSummary:
      type: object
      additionalProperties: false
      required: [venueId, name, city, country]
      properties:
        venueId:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
        city:
          type: string
          minLength: 1
        country:
          type: string
          minLength: 1
    VenueCreate:
      type: object
      additionalProperties: false
      required: [name, city, country]
      properties:
        name:
          type: string
          minLength: 1
        street:
          type: string
        zip:
          type: string
        city:
          type: string
          minLength: 1
        country:
          type: string
          minLength: 1
        latitude:
          type: number
        longitude:
          type: number
        notes:
          type: string
    VenueUpdate:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
        street:
          type: string
        zip:
          type: string
        city:
          type: string
          minLength: 1
        country:
          type: string
          minLength: 1
        latitude:
          type: number
        longitude:
          type: number
        notes:
          type: string
    Meet:
      type: object
      additionalProperties: false
      required:
        - meetId
        - title
        - startAt
        - status
        - venueId
        - organizerUserId
        - createdAt
      properties:
        meetId:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
        description:
          type: string
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [PLANNED, CANCELLED, DONE]
        venueId:
          type: string
          format: uuid
        organizerUserId:
          type: string
          format: uuid
          description: UserId from Registration-Service (no DB FK here)
        maxParticipants:
          type: integer
          minimum: 0
        maxVisitors:
          type: integer
          minimum: 0
        createdAt:
          type: string
          format: date-time
    MeetResponse:
      allOf:
        - $ref: "#/components/schemas/Meet"
        - type: object
          additionalProperties: false
          required: [venue]
          properties:
            venue:
              $ref: "#/components/schemas/VenueSummary"
    MeetCreate:
      type: object
      additionalProperties: false
      required: [title, startAt, venueId, organizerUserId]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        venueId:
          type: string
          format: uuid
        organizerUserId:
          type: string
          format: uuid
        maxParticipants:
          type: integer
          minimum: 0
        maxVisitors:
          type: integer
          minimum: 0
    MeetUpdate:
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [PLANNED, CANCELLED, DONE]
        venueId:
          type: string
          format: uuid
        maxParticipants:
          type: integer
          minimum: 0
        maxVisitors:
          type: integer
          minimum: 0
    User:
      type: object
      additionalProperties: false
      required: [userId, displayName, email, createdAt]
      properties:
        userId:
          type: string
          format: uuid
        displayName:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
    UserCreate:
      type: object
      additionalProperties: false
      required: [displayName, email]
      properties:
        displayName:
          type: string
          minLength: 1
        email:
          type: string
          format: email
    UserUpdate:
      type: object
      additionalProperties: false
      properties:
        displayName:
          type: string
          minLength: 1
        email:
          type: string
          format: email
    Vehicle:
      type: object
      additionalProperties: false
      required: [vehicleId, userId, make, model]
      properties:
        vehicleId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        make:
          type: string
          minLength: 1
        model:
          type: string
          minLength: 1
        year:
          type: integer
          minimum: 1900
        color:
          type: string
        powerHp:
          type: integer
          minimum: 0
        plateOptional:
          type: string
        notes:
          type: string
    VehicleCreate:
      type: object
      additionalProperties: false
      required: [make, model]
      properties:
        make:
          type: string
          minLength: 1
        model:
          type: string
          minLength: 1
        year:
          type: integer
          minimum: 1900
        color:
          type: string
        powerHp:
          type: integer
          minimum: 0
        plateOptional:
          type: string
        notes:
          type: string
    VehicleUpdate:
      type: object
      additionalProperties: false
      properties:
        make:
          type: string
          minLength: 1
        model:
          type: string
          minLength: 1
        year:
          type: integer
          minimum: 1900
        color:
          type: string
        powerHp:
          type: integer
          minimum: 0
        plateOptional:
          type: string
        notes:
          type: string
    VehicleCreateWithUser:
      type: object
      additionalProperties: false
      required: [userId, make, model]
      properties:
        userId:
          type: string
          format: uuid
        make:
          type: string
          minLength: 1
        model:
          type: string
          minLength: 1
        year:
          type: integer
          minimum: 1900
        color:
          type: string
        powerHp:
          type: integer
          minimum: 0
        plateOptional:
          type: string
        notes:
          type: string
    Registration:
      type: object
      additionalProperties: false
      required: [registrationId, meetId, userId, role, status, createdAt]
      properties:
        registrationId:
          type: string
          format: uuid
        meetId:
          type: string
          format: uuid
          description: MeetId from Meet-Service (no DB FK here)
        userId:
          type: string
          format: uuid
        vehicleId:
          type: [string, "null"]
          format: uuid
          description: Nullable for VISITOR registrations
        role:
          type: string
          enum: [PARTICIPANT, VISITOR]
        status:
          type: string
          enum: [CONFIRMED, CANCELLED]
        createdAt:
          type: string
          format: date-time
        checkInAt:
          type: [string, "null"]
          format: date-time
    RegistrationCreate:
      type: object
      additionalProperties: false
      required: [meetId, userId, role]
      properties:
        meetId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        role:
          type: string
          enum: [PARTICIPANT, VISITOR]
        vehicleId:
          type: [string, "null"]
          format: uuid
          description: Required if role=PARTICIPANT, must be null if VISITOR
    RegistrationUpdate:
      type: object
      additionalProperties: false
      properties:
        vehicleId:
          type: [string, "null"]
          format: uuid
          description: Nullable for VISITOR registrations
        role:
          type: string
          enum: [PARTICIPANT, VISITOR]
        status:
          type: string
          enum: [CONFIRMED, CANCELLED]
        checkInAt:
          type: [string, "null"]
          format: date-time

    RegistrationCount:
      type: object
      additionalProperties: false
      required: [meetId, total]
      properties:
        meetId:
          type: string
          format: uuid
        total:
          type: integer
          minimum: 0
        participants:
          type: integer
          minimum: 0
        visitors:
          type: integer
          minimum: 0